Test de integración y fixes
-Nombres constantes
-CircuitEnd2End
-fix database

+++ src/main/java/com/uah/f1backend/repository/CircuitModelRepository.java
@@ -1,13 +1,10 @@
 package com.uah.f1backend.repository;
 
-
 import com.uah.f1backend.model.CircuitModel;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
-
-public interface CircuitModelRepository extends JpaRepository<CircuitModel, Long> {
-    //Find a circuit by its name
+public interface CircuitModelRepository extends JpaRepository<CircuitModel, Integer> {
+    // Find a circuit by its name
     Optional<CircuitModel> findByName(String name);
 }
